Extract stack instructions to instructions_stack.go module
